[Refactor][envtest] Centralize all helpers in envtest for better DX
