+++ ray-operator/controllers/ray/rayservice_controller_test.go
@@ -675,136 +675,3 @@ var _ = Context("Inside the default namespace", func() {
 		})
 	})
 })
-
-func prepareFakeRayDashboardClient() *utils.FakeRayDashboardClient {
-	client := &utils.FakeRayDashboardClient{}
-
-	healthyStatus := generateServeStatus(rayv1.DeploymentStatusEnum.HEALTHY, rayv1.ApplicationStatusEnum.RUNNING)
-	client.SetMultiApplicationStatuses(map[string]*utils.ServeApplicationStatus{"app": &healthyStatus})
-
-	return client
-}
-
-func generateServeStatus(deploymentStatus string, applicationStatus string) utils.ServeApplicationStatus {
-	return utils.ServeApplicationStatus{
-		Status: applicationStatus,
-		Deployments: map[string]utils.ServeDeploymentStatus{
-			"shallow": {
-				Name:    "shallow",
-				Status:  deploymentStatus,
-				Message: "",
-			},
-			"deep": {
-				Name:    "deep",
-				Status:  deploymentStatus,
-				Message: "",
-			},
-			"one": {
-				Name:    "one",
-				Status:  deploymentStatus,
-				Message: "",
-			},
-		},
-	}
-}
-
-func getRayClusterNameFunc(ctx context.Context, rayService *rayv1.RayService) func() (string, error) {
-	return func() (string, error) {
-		if err := k8sClient.Get(ctx, client.ObjectKey{Name: rayService.Name, Namespace: "default"}, rayService); err != nil {
-			return "", err
-		}
-		return rayService.Status.ActiveServiceStatus.RayClusterName, nil
-	}
-}
-
-func getActiveRayClusterWorkerGroupSpecsFunc(ctx context.Context, rayService *rayv1.RayService) func() ([]rayv1.WorkerGroupSpec, error) {
-	return func() ([]rayv1.WorkerGroupSpec, error) {
-		if err := k8sClient.Get(ctx, client.ObjectKey{Name: rayService.Name, Namespace: "default"}, rayService); err != nil {
-			return nil, err
-		}
-		rayCluster := &rayv1.RayCluster{}
-		if err := k8sClient.Get(ctx, client.ObjectKey{Name: rayService.Status.ActiveServiceStatus.RayClusterName, Namespace: "default"}, rayCluster); err != nil {
-			return nil, err
-		}
-		return rayCluster.Spec.WorkerGroupSpecs, nil
-	}
-}
-
-func getPreparingRayClusterNameFunc(ctx context.Context, rayService *rayv1.RayService) func() (string, error) {
-	return func() (string, error) {
-		if err := k8sClient.Get(ctx, client.ObjectKey{Name: rayService.Name, Namespace: "default"}, rayService); err != nil {
-			return "", err
-		}
-		return rayService.Status.PendingServiceStatus.RayClusterName, nil
-	}
-}
-
-func getPendingRayClusterWorkerGroupSpecsFunc(ctx context.Context, rayService *rayv1.RayService) func() ([]rayv1.WorkerGroupSpec, error) {
-	return func() ([]rayv1.WorkerGroupSpec, error) {
-		if err := k8sClient.Get(ctx, client.ObjectKey{Name: rayService.Name, Namespace: "default"}, rayService); err != nil {
-			return nil, err
-		}
-		rayCluster := &rayv1.RayCluster{}
-		if err := k8sClient.Get(ctx, client.ObjectKey{Name: rayService.Status.PendingServiceStatus.RayClusterName, Namespace: "default"}, rayCluster); err != nil {
-			return nil, err
-		}
-		return rayCluster.Spec.WorkerGroupSpecs, nil
-	}
-}
-
-func checkServiceHealth(ctx context.Context, rayService *rayv1.RayService) func() (bool, error) {
-	return func() (bool, error) {
-		if err := k8sClient.Get(ctx, client.ObjectKey{Name: rayService.Name, Namespace: rayService.Namespace}, rayService); err != nil {
-			return false, err
-		}
-
-		for _, appStatus := range rayService.Status.ActiveServiceStatus.Applications {
-			if appStatus.Status != rayv1.ApplicationStatusEnum.RUNNING {
-				return false, nil
-			}
-			for _, deploymentStatus := range appStatus.Deployments {
-				if deploymentStatus.Status != rayv1.DeploymentStatusEnum.HEALTHY {
-					return false, nil
-				}
-			}
-		}
-
-		return true, nil
-	}
-}
-
-// Update the status of the head Pod to Running.
-// We need to manually update Pod statuses otherwise they'll always be Pending.
-// envtest doesn't create a full K8s cluster. It's only the control plane.
-// There's no container runtime or any other K8s controllers.
-// So Pods are created, but no controller updates them from Pending to Running.
-// See https://book.kubebuilder.io/reference/envtest.html for more details.
-func updateHeadPodToRunningAndReady(ctx context.Context, rayClusterName string) {
-	headPods := corev1.PodList{}
-	headFilterLabels := client.MatchingLabels{
-		utils.RayClusterLabelKey:  rayClusterName,
-		utils.RayNodeTypeLabelKey: string(rayv1.HeadNode),
-	}
-
-	Eventually(
-		listResourceFunc(ctx, &headPods, headFilterLabels, &client.ListOptions{Namespace: "default"}),
-		time.Second*15, time.Millisecond*500).Should(Equal(1), "Head pod list should have only 1 Pod = %v", headPods.Items)
-
-	headPod := headPods.Items[0]
-	headPod.Status.Phase = corev1.PodRunning
-	headPod.Status.Conditions = []corev1.PodCondition{
-		{
-			Type:   corev1.PodReady,
-			Status: corev1.ConditionTrue,
-		},
-	}
-	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		return k8sClient.Status().Update(ctx, &headPod)
-	})
-	Expect(err).NotTo(HaveOccurred(), "Failed to update head Pod status to PodRunning")
-
-	// Make sure the head Pod is updated.
-	Eventually(
-		isAllPodsRunning(ctx, headPods, headFilterLabels, "default"),
-		time.Second*15, time.Millisecond*500).Should(BeTrue(), "Head Pod should be running: %v", headPods.Items)
-}
